+++ gui/iode_objs/views/variables_view.h
@@ -9,7 +9,7 @@
 #include "iode_objs/delegates/variables_delegate.h"
 #include "iode_objs/new/add_variable.h"
 #include "plot/plot_vars.h"
-#include "custom_widgets/print_file_dialog.h"
+#include "print/print_file_dialog.h"
 #include "menu/print_graph/graph_variables.h"
 #ifndef TEMPLATE_NUMERICAL_VIEW_HEADER
 #include "numerical_view.h"
